refactor: optimize technologies data
