+++ lib/constants.ts
@@ -1,0 +1,1 @@
+export const displayName = ["S", "h", "i", "n", "j", "i", "t", "h"];
