+++ lib/data/about.json
@@ -1,0 +1,5 @@
+{
+  "title": "Web Developer",
+  "caption": "Based in Kerala, India",
+  "description": "Who specializes in creating pixel-perfect, engaging, and accessible digital experiences. Currently, I'm concentrating on developing useful web projects and learning new things."
+}
