+++ lib/data/index.ts
@@ -1,0 +1,7 @@
+import aboutJson from "./about.json";
+
+const data = {
+  about: aboutJson as DataAbout,
+};
+
+export default data;
