+++ lib/data/technologies.tsx
@@ -1,0 +1,157 @@
+import {
+  IconBrandAws,
+  IconBrandJavascript,
+  IconBrandNextjs,
+  IconBrandReact,
+  IconBrandReactNative,
+  IconBrandTypescript,
+  IconBrandYarn,
+  IconBrandRedux,
+} from "@tabler/icons-react";
+import {
+  SiBootstrap,
+  SiC,
+  SiCloudflareworkers,
+  SiDjango,
+  SiDocker,
+  SiElectron,
+  SiExpo,
+  SiExpress,
+  SiMantine,
+  SiMarkdown,
+  SiMui,
+  SiNetlify,
+  SiNodedotjs,
+  SiNpm,
+  SiPnpm,
+  SiPython,
+  SiRender,
+  SiSocketdotio,
+  SiTailwindcss,
+  SiTauri,
+} from "@icons-pack/react-simple-icons";
+
+export const technologies: Technology[] = [
+  {
+    name: "AWS",
+    icon: IconBrandAws,
+  },
+  {
+    name: "Bootstrap",
+    icon: SiBootstrap,
+  },
+  {
+    name: "C",
+    icon: SiC,
+  },
+  {
+    name: "Workers",
+    icon: SiCloudflareworkers,
+  },
+  {
+    name: "Django",
+    icon: SiDjango,
+  },
+  {
+    name: "Docker",
+    icon: SiDocker,
+  },
+  {
+    name: "Electron",
+    icon: SiElectron,
+  },
+
+  {
+    name: "Expo",
+    icon: SiExpo,
+  },
+
+  {
+    name: "Express",
+    icon: SiExpress,
+  },
+
+  {
+    name: "JavaScript",
+    icon: IconBrandJavascript,
+  },
+  {
+    name: "Mantine",
+    icon: SiMantine,
+  },
+
+  {
+    name: "Markdown",
+    icon: SiMarkdown,
+  },
+  {
+    name: "Material UI",
+    icon: SiMui,
+  },
+
+  {
+    name: "Netlify",
+    icon: SiNetlify,
+  },
+
+  {
+    name: "Next.js",
+    icon: IconBrandNextjs,
+  },
+  {
+    name: "Node",
+    icon: SiNodedotjs,
+  },
+  {
+    name: "NPM",
+    icon: SiNpm,
+  },
+  {
+    name: "PNPM",
+    icon: SiPnpm,
+  },
+
+  {
+    name: "Python",
+    icon: SiPython,
+  },
+
+  {
+    name: "React",
+    icon: IconBrandReact,
+  },
+  {
+    name: "React Native",
+    icon: IconBrandReactNative,
+  },
+  {
+    name: "Render",
+    icon: SiRender,
+  },
+  {
+    name: "Redux",
+    icon: IconBrandRedux,
+  },
+  {
+    name: "Socket IO",
+    icon: SiSocketdotio,
+  },
+  {
+    name: "Tailwind",
+    icon: SiTailwindcss,
+  },
+
+  {
+    name: "Tauri",
+    icon: SiTauri,
+  },
+  {
+    name: "TypeScript",
+    icon: IconBrandTypescript,
+  },
+
+  {
+    name: "Yarn",
+    icon: IconBrandYarn,
+  },
+];
