+++ lib/data/types.d.ts
@@ -1,0 +1,10 @@
+type DataAbout = {
+  title: string;
+  caption: string;
+  description: string;
+};
+
+type Technology = {
+  icon: React.ElementType;
+  name: string;
+};
